backend/requirements.txt
Flask==2.3.3
pytesseract==0.3.10
APScheduler==3.10.4
Flask-Mail==0.9.1
Werkzeug==2.3.7
Pillow==10.0.1
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0